requirements.txt
# MCP Framework
mcp[cli]>=1.0.0

# LLM
anthropic>=0.40.0

# HTTP Client
httpx>=0.27.0

# Validation
pydantic>=2.0.0

# Environment
python-dotenv>=1.0.0

# Document Export
python-docx>=1.1.0

# Frontend
streamlit>=1.38.0
